Add config file for enabling LGTM analysis
lgtm.com is already picking up qtbase's mirror repo on github, trying to
analyze it. Because it lacks build instructions, lgtm.com fails to analyze the
C++ code, and only presents flaws in the little python and javascript code
that is in the repo. With the file in this commit, it should be able to
analyze the C++ code too.

Change-Id: If558d26800e76c1b8590ed1aaf28adb6c0d6e188
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>